change table.js to party.js
Updates all references to `table` to be `party`

lines 35-39: fixes problem where updating Party information will remove all information from the document (except for the provided information)
